+++ Runtime/Animation/UIAnimChannel.cs
@@ -1,0 +1,10 @@
+namespace KitforgeLabs.MobileUIKit.Animation
+{
+    public enum UIAnimChannel
+    {
+        Scale = 0,
+        Fade = 1,
+        Position = 2,
+        Rotation = 3
+    }
+}