+++ Runtime/Toast/UIToast.cs
@@ -1,0 +1,9 @@
+namespace KitforgeLabs.MobileUIKit.Toast
+{
+    public abstract class UIToast<TData> : UIToastBase
+    {
+        public abstract void Bind(TData data);
+
+        protected internal sealed override void BindUntyped(object data) => Bind((TData)data);
+    }
+}